@@ -1,20 +1,16 @@
-[tool.poetry]
+[project]
 name = "lager"
-version = "2022.03.13"
+version = "2024.11.30"
 description = "Stupid Simple Logging 🍺"
-authors = ["Vince Forgione <vince.4gione@icloud.com>"]
+authors = [{ name = "Vince Forgione", email = "vince.4gione@icloud.com" }]
 license = "MIT"
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = []
 
-[tool.poetry.dependencies]
-python = ">=3.8,<4.0"
-
-[tool.poetry.dev-dependencies]
-black = "^22.1.0"
-isort = "^5.10.1"
-pylint = "^2.12.2"
-pytest = "^7.0.1"
-pytest-cov = "^3.0.0"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
+[dependency-groups]
+dev = [
+    "pyright>=1.1.389",
+    "pytest>=8.3.3",
+    "ruff>=0.8.1",
+]
